Update CMake presets and readme (#3523)
* Update CMake presets and readme

* Switch to Visual Studio generator to fix vcpkg/cmake x86-32 issues

* Add Windows x64 ci

* Add casts on size_t where required

+++ src/OpenLoco/src/GameCommands/Industries/CreateIndustry.cpp
@@ -373,7 +373,7 @@ namespace OpenLoco::GameCommands
             return Colour::black;
         }
         // Note: Don't optimise for size 1 as randNext required to prevent divergence
-        return availableColours[prng.randNext(availableColours.size() - 1)];
+        return availableColours[prng.randNext(static_cast<int32_t>(availableColours.size()) - 1)];
     }
 
     struct RangeAndMinDistance
